airscan-zeroconf: Accept direct device specification
Passing a string in the format "protocol:name:url" to sane_open() will
attempt to directly connect to a device named `name' at `url' with
`protocol' as through a line like this had been read from the config
file:

name = url, protocol

Direct device names can be distinguished from discovered and statically
configured identifiers because they start with a protocol name instead
of e0/w1/etc.  Because of that, the logic will attempt to parse a direct
name before starting discovery or checking the config.  If the
identifier can be parsed, the other mechanisms are skipped.