Add AF_UNIX support to HTTP connections
Allow the URL for a device to be specified as unix://socket_name/... in
addition to the existing http:// URLs.  Also add a new socket_dir option
to the [options] section of airscan.conf.

When a unix:// URL is encountered, the sockdir_dir and socket_name
values are combined to form a full path to a local AF_UNIX socket.
Instead of attempting a TCP connection to the device, all HTTP requests
for that device are sent through the local socket.  The socket file
itself must be created by some external program that knows how to
communicate with the scanner, such as the Chrome OS ippusb_bridge that
forwards through an IPP-USB tunnel.

The URL is parsed with the same parser as http:// URLs, so the socket
name is currently restricted to valid DNS hostname characters.

Fixes alexpevzner/sane-airscan-unstable#11