Suppress clang false positive warning
clang 11 adds a new on-by-default diagnostic for casting a void pointer
to a smaller integer type (-Wvoid-pointer-to-enum-cast).  This causes a
warning in mdns_avahi_browser_callback when casting the userdata void
pointer to an MDNS_SERVICE value.  The warning is a false positive
because userdata is always passed as a value from the MDNS_SERVICE enum,
so we can cast it through intptr_t to eliminate the warning.