WSDD: WS-Discovery reliability improved for Pantum devices
WS-Discovery implementation contains a trade-off between reliability
and response time.

To discover devices it sends Probe requests for some time in a hope
that all devices will be able to respond within this time limit. The
longer this time, the more reliable the search but the more time it
takes. And we cannot make this time dynamically adaptive, because we
don't have a priori knowledge about devices to be found.

The currently chosen discovery time, 2500 milliseconds (WSDD_DISCOVERY_TIME)
looks the reasonable trade-off in most cases. However, for many Pantum
devices it is often not enough for reliable discovery.

The implemented solution was to increase this time up to 5000 milliseconds,
if potentially problematic (Pantum for now) devices were discovered on a
network, using DNS-SD.

The list of problematic models may grow in a future, for now it contains
only Pantums, but implementation allows it to be easily extended.