MDNS: fix discovery if device misses IPv6 address or so (see #326)
This change adds ZEROCONF_PUBLISH_DELAY support for MDNS. The idea is
when at leas IP address is discovered, don't wait forever until all
related AvahiServiceResolver are finished or timed out, but limit
the wait time by ZEROCONF_PUBLISH_DELAY parameter (1000 milliseconds
for now).

Without this measure, if IPv6 address of device is not available at all,
we will know that only when AvahiServiceResolver times out, which is a
bit too late.