WSD: fixed ADF duplex on Epson Workforce WF-3520 (see #339)
On devices, like Epson Workforce WF-3520, which report different scan
region size in ADF duplex and simplex modes, we failed to detect and
report ADF duplex mode.

It happened only in WSD and happened because WSD uses unusual model
of reporting device capabilities: it reports separately capabilities
for the "ADF front page" scan and separately for the "ADF back page
scan", while SANE assumes in the duplex mode these capabilities are
the same.

The problem was in bad interference between algorithm that figures
out the common subset of these capabilities (implemented in the
devcaps_source_merge() function) and workaround for Kyocera, which
(the workaround) disables scan region clipping by setting minimal
region width and height to be equal to maximal. When two sources offer
slightly different geometry and min forced to be max, there is obviously
no common range for these two sources.