eSCL: added quirk_adf_max_pages
Some printers fail to properly scan the entire ADF load. In particular:

  * Ricoh MP 501 scans properly few pages, but fails on a full ADF
    load (Issue #356)
  * EPSON WF-2930 consumes all pages from the ADF, but returns image
    only for the first page (Issue #357)

This quirk was added for experimenting with finding workaround for
these problems. If set, it adds scan:NumberOfPages parameter into
the eSCL ScanJobs request.

Note, not all devises properly support this parameter, so users feedback
is required to test this stuff on their hardware.

If these printers will properly handle scan:NumberOfPages, the workaround
is obvious: limit ADF scan with amount of pages that printer can properly
handle, then (automatically) restart the scan, until ADF is empty. It may
slightly slow down the scanning, but at least it will work correctly.

For now, scan:NumberOfPages is hardcoded to "1" for both devices with a
goal to verify the principal possibility of this approach. Other devices
are not affected.